Add my clang-format for INFOMOV/MAGR
This config was made for these courses and tries to mimmic Jacco's style
as close as possible, while still being sensible.
The only difference is pointer alignment. This project uses left-style
everywhere, so there's not much point changing that after the fact and
generating tons of changes.

However, the recently added Vulkan RT core by Mèir has been made with
that original .clang-format, using the (IMO) correct
PointerAlignment: Right. This version is added in the Vulkan_RT
subdirectory as clang formatters always recursively walk up the
directory (relative to the source file) to find the right .clang-format
file.