CMake: Create explicit dependency from apps on RenderCores in macro
This dependency is necessary when running the startup project in Visual
Studio. It is expected to build RenderCores and dependencies on changes
when developing and debugging.
Note that the dependency is NOT created on non-windows targets, for the
time being.

TODO: Move macro to separate file, and include as much obvious duplicate
functionality.
Same for RenderCores, such as linking against RenderSystem and setting
COREDLL_EXPORTS.