CMake: Prefer Ninja build system over Unix Makefiles
The generated build system is better able to use all processing cores,
which is vitally important with a single long-running compilation on
host_meshloaders.cpp with the massive tinygltf and tinyobj loaders.
The build rules for Ninja "understand" that compiling source files for
other libraries (usually RenderCores) do not depend on another library
(libRenderSystem in this case) to be completed. This dependency is only
necessary at linking stage. This allows a more constant use of all
processor cores, and the build finishing in 5/6th to 3/4th of the time
it would take with Unix Makefiles. This is likely a limitation of the
CMake Makefiles generator, since there is nothing preventing Makefiles
from having the same optimized rules.