RenderSystem: Remove dlopening of unknown file
Trying to load a file without lib and .so doesn't make sense in this
context. It is not the current directory (prepend ./ for that), which is
controlled by the RPATH anyway. In fact, RPATH (or ld.so paths) _must_
define this since prefixing `./` would only apply to the dlopened solib,
all their dependencies will still be searched merely where ld.so is
taught to look:
https://gitlab.kitware.com/cmake/community/-/wikis/doc/cmake/RPATH-handling