CMake: RenderSystem: Search for cores in $<CONFIG> subdir on multiconfig
When building in a multiconfig environment (eg. Visual Studio) all DLLs
are stored in a subdirectory matching the selected build config as to
not confuse files of different configurations (this config can change on
the fly).
Make sure these are properly selected and that no subdirectory is
entered on MSVC CMake builds from the cmdline/VSCode (single-config
CMake doesn't generate a debug/ nor release/ subdirectory like the usual
Visual Studio solution builds).