lib: Vulkan: Update Windows SDK to 1.2.131.2
Update Vulkan headers to the last version that does not hardwire NV
raytracing to the new KHR extension (which would be 1.2.135.0). This
requires a beta driver that is currently not installed on most systems.
A bump to that SDK, as well as a refactor to utilize the KHR extension
and drop (now legacy) NV calls.

Note that shaderc _is_ updated to 1.2.135.0 because the headers on
1.2.131.2 are incomplete. shaderc_combined.lib is dropped because it
weighs in at over 100MB which is not accepted by GitHub, and the file is
not used anyway.